added variable name resolution

+++ src/SurpherCallable.cpp
@@ -9,7 +9,7 @@ SurpherFunction::SurpherFunction(std::shared_ptr<Function> declaration, std::sha
 
 std::any SurpherFunction::call(Interpreter &interpreter, const std::vector<std::any> &arguments) {
     std::shared_ptr<Environment> environment = std::make_shared<Environment>(closure);
-    for (uint32_t i = 0; i < declaration->params.size(); i++) {
+    for (size_t i = 0; i < declaration->params.size(); i++) {
         environment->define(declaration->params[i].lexeme, arguments[i]);
     }
     try {
